Moving affine transform to a separate file
